requirements.txt
esptool==2.6
pyserial==3.5